requirements.txt
accelerate==1.0.1
bitsandbytes==0.44.1
comet_ml==3.46.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
ftfy
future
gdown
gradio
gradio-client
imageio
iopath
ipykernel
loralib==0.1.2
lxml==5.3.0
matplotlib==3.9.2
ninja
nltk==3.9.1
numpy
omegaconf==2.3.0
openai==1.56.0
opencv-python==4.10.0.84
Pillow==11.0.0
POT==0.9.4
pycocotools==2.0.8
pydantic==2.9.2
pytorch_lightning==2.3.3
PyYAML==6.0.1
scikit-image
scikit-learn==1.2.2
scipy==1.14.1
seaborn
tensorboardX==2.6.2.2
timm==0.6.13
torchmetrics==1.4.2
torchshow
tqdm==4.66.4
transformers==4.46.2
ttach==0.0.3
umap-learn
